+++ cmake/modules/UseInkscape.cmake
@@ -22,13 +22,6 @@
 #
 #    TODO Switch to named arguments!
 #
-#
-# .. cmake_function:: inkscape_generate_eps_from_svg
-#
-#    .. deprecated:: 2.7
-#       Switch to PNG instead of EPS and use
-#       inkscape_generate_png_from_svg.
-#
 
 include(CMakeParseArguments)
 
@@ -48,31 +41,6 @@ function(inkscape_generate_png_from_svg)
     string(REGEX REPLACE "\\.[a-zA-Z]+" ".svg" input ${pic})
     execute_process(
       COMMAND ${INKSCAPE} -z --export-dpi=${INKSCAPE_DPI} -e ${pic} ${CMAKE_CURRENT_SOURCE_DIR}/${input}
-      WORKING_DIRECTORY  ${INKSCAPE_OUTPUT_DIR})
-  endforeach()
-endfunction()
-
-function(inkscape_generate_eps_from_svg)
-  message(AUTHOR_WARNING "inkscape_generate_eps_from_svg is deprecated and will be removed after Dune 2.7.")
-  cmake_parse_arguments(INKSCAPE "" "INPUT_DIR;OUTPUT_DIR;DPI" "" ${ARGN})
-  if(NOT INKSCAPE_INPUT_DIR)
-    set(INKSCAPE_INPUT_DIR ${CMAKE_CURRENT_SOURCE_DIR})
-  endif()
-  if(NOT INKSCAPE_INPUT_DIR)
-    set(INKSCAPE_OUTPUT_DIR ${CMAKE_CURRENT_BINARY_DIR})
-  endif()
-
-  foreach(_pic ${INKSCAPE_UNPARSED_ARGUMENTS})
-    string(REGEX REPLACE "\\.[a-zA-Z]+" ".png" input "${_pic}")
-    string(REGEX REPLACE "\\.[a-zA-Z]+" ".svg" svginput "${_pic}")
-
-    add_custom_target(${input}
-      COMMAND ${INKSCAPE} -z --export-dpi=${INKSCAPE_DPI} -e ${input} ${CMAKE_CURRENT_SOURCE_DIR}/${svginput}
-      COMMENT "Generating ${INKSCAPE_OUTPUT_DIR}/${svginput} from ${CMAKE_CURRENT_SOURCE_DIR}/${input}")
-    add_custom_command(OUTPUT ${_pic}
-      COMMAND ${CONVERT} ${INKSCAPE_OUTPUT_DIR}/${input} EPS:${_pic}
-      DEPENDS ${input}
-      COMMENT "Converting ${INKSCAPE_OUTPUT_DIR}/${input} to ${INKSCAPE_OUTPUT_DIR}/${_pic}"
       WORKING_DIRECTORY  ${INKSCAPE_OUTPUT_DIR})
   endforeach()
 endfunction()